Moving `Pred` to core
